[socket-nif] nosup expection of win32 and type(s) replacements
The nif callback functions (nif_open) now instead cause an
'nosup' exception if called (instead of badarg).

The basic type uint16_t, uint32_t and int32_t (C99) replaced
"own" (that is, defined by  "us") types Uint16, Uint32 and
Sint32.
The point of this is that our Windows build system seems to
be a bit lacking when it comes to types...

Removed "some stuff" that was if-defed. Different solution
when win32 support for sockets has been improved.

Make sure the socket_*.c util modules are not included
in the building for windows.

OTP-15526