Make sure that beam_ssa_opt optimizes all functions
The function `get_call_order_po/2` did not always include all
functions in a module. The method it used was to first find
all leave functions (functions not calling any other function),
and then find all others functions that called any of the leave
functions either directly or indirectly.

Functions that did not call a leave function (directly or indirectly)
would not be included in the list of functions to optimize.

Reimplement `get_call_order_po/2` to use the standard algorithm
for constructing a list of nodes in reverse postorder and then
reverse that list.