beam_ssa_opt: Do local CSE of get_tuple_element instructions
For some reason, a `get_tuple_element` instruction was not deemed
suitble for local common sub expression elimination.

It turns out that enabling CSE for `get_tuple_element` is benefical.
It will also be even more benefical in a future commit where some
of the optimizations in `sys_core_fold` are removed.