[logger] Make sure log file is re-opened with configured file options
Earlier, if the log file had to be re-opened, e.g. due to it being
renamed or removed, it would always be opened with the default file
options, even if other options were set in the configuration.

Now, the configured options are used, except that 'append' is always
added and 'exclusive' is always removed.