otp_build: Add -t option for building a tiny system
Add the `-t` option to `otp_build` for building a tiny system. The
tiny system can be used for updating the primary bootstrap or
the preloaded modules in erts.

On my computer, `./otp_build setup -t` finished the build of the
tiny system in 1 min 50 seconds. The small build (`-s` option or
no option) finished in 2 min 55 seconds.