beam_validator: Track types by value rather than by register
This is a rather subtle but important distinction. While tracking
types on a per-register basis is fairly effective, it forces us to
track which registers alias each other, and makes it tricky to infer
types over large blocks of code as instruction arguments may have
been clobbered between definition and inference.

Tracking types on a per-value basis makes us immune to these
problems.