Warn on duplicate apps or duplicate keys in config
Prior to this patch, the behaviour was unspecified,
and duplicate keys and duplicate apps would behave
different depending on the amount of config, the name
of the config files, and how those configs would be
listed.

This commit warns on such scenario, so we can raise
errors in the future.

    $ erl -config dupkeys.config -s erlang halt
    2019-02-27 11:16:02.653100 application: ~ts; duplicate parameter: ~ts~n
        kernel
        key1
    =WARNING REPORT==== 27-Feb-2019::11:16:02.653100 ===
    application: kernel; duplicate parameter: key1

    $ erl -config dupapps.config -s erlang halt
    2019-02-27 11:16:02.653100 duplicate application config: ~ts
        kernel
    =WARNING REPORT==== 27-Feb-2019::11:16:02.653100 ===
    duplicate application config: kernel