Rewords one sentence in common_test documentation
Clarify and avoid misunderstandings. In the previous sentence,
it is written that "the test case is the smallest unit...". To me
it was confusing when I read the following sentence: "Subsets of
test cases, ...". The key part of the paragraph is that test cases
can be grouped, not that they are subsets.