Proposal. Mnesia. Introduce a new lock kind: none, to skip locking.
The new lock kind is supported in write,delete and delete_object

The lock kind will skip locking for the operations.
It's an unsafe option and should only be used if there are other
locks protecting the record, for example global locks.

The reason behind this change is to improve performance in cases
like cleanup, where many records may be accessed at once, but still
there may be more than one transaction modifying different parts
of the table. In this cases it's convenient to use global locks to
separate sections of the database affected.
If locks are not table-wise, each write/delete operation will acquire
a separate lock, which degrades performance and may not be necessary,
since the transaction is already protected by the global lock.

This happens in RabbitMQ when a queue is deleted. The queue delete
transaction will cleanup all the bindings for that queue. Same thing
when an exchange is deleted.
Bindings in this case are scoped by the queue or exchange, but there
may be many of them, all trying to acquire a separate cluster-wide
write lock.
RabbitMQ uses a global lock per queue/exchange when deleting to prevent
other operations on this queue/exchange bindings. This makes the separate
bindings locks useless.