inets: Fix handling of erl_script_timeout
When httpd was started with 'erl_script_timeout', the value of the
option was converted to milliseconds before storage. Subsequent
calls to httpd:info/1 returned the input value multiplied by 1000.

This change fixes the handing of erl_script_timeout by storing the
timeout in seconds and converting it to milliseconds before usage.

Change-Id: Ic308d83b59ad0884e053f096f995754c89adcbf4