[socket] Make use of the new select (read|write) functions
Make use of the new select functions; enif_select_[read|write],
for read and write select. These functions allows us to construct
the select message ourseves:

    {'$socket', Socket, select, Ref}

This is in preparations for when we introduce the 'nowait'
(or something similar) value for the timeout argument (in accept,
read and write funcions).

It also solves (we hope) the term leakage problems (it was difficult
to free the environment when there was only one/socket).

OTP-15496