Implement inet_crypto_dist benchmark example
This is a simplistic distribution protocol module for
encrypted distribution using a shared secret and only
the crypto module, and AEAD ciphers.

It is intended to be a benchmark reference for how fast
encrypted distribution may get.

How secure it would be and also how useful, for actual production use,
remains to be investigated.