seq_trace: Inherit trace token on spawn
Trace tokens can be lost when a process delegates message sending
to a child process, which is pretty surprising and limits the
usefulness of seq tracing. One example of this is gen_statem:call/4
which uses a child process to implement timeouts without the risk
of a late message arriving to the caller.

This commit attempts to remedy this by propagating the trace token
to spawned processes, and adds optional tracing for process
spawning as well.