Optimize initialization of Y registers for the JIT
The `allocate_zero` and `allocate_heap_zero` instructions allocate a
new initialized stack frame. The compiler will emit those instructions
when more than half of Y registers are not naturally initialized by
saving X registers to them before the first instruction that causes a
GC or the stack to be scanned.

Those instructions make sense for the interpreted BEAM since they are
cheaper than allocating an uninitialized stack frame (with `allocate`)
and using a sequence of `init` instruction to initialize the Y
registers that are not otherwise initialized. They do not make sense
for the JIT because it makes it impossible to avoid to initializing
some Y registers twice.

This commit introduces the `init_yregs` instruction that initializes
the Y registers given in its list operand.

As an example, consider allocating a stack frame of six Y registers,
two of which are initialized by saving X registers to them. Here is
how the instructions could look like before this commit:

    {allocate_zero,6,2}.
    {move,{x,0},{y,0}}.
    {move,{x,1},{y,1}}.
    {call,...}.

With `init_yregs` instruction available, the compiler will emit
the following instructions:

    {allocate,6,2}.
    {init_yregs,{list,[{y,2},{y,3},{y,4},{y,5}]}}.
    {move,{x,0},{y,0}}.
    {move,{x,1},{y,1}}.
    {call,...}.

The JIT can emit an efficient native-code sequence to initialize
`{y,2}` through `{y,5}`. For the interpreter, the loader will
translate the `allocate` and `init_yregs` back to an `allocate_zero`
instruction.

The `init_yregs` instruction will also be used instead of a
sequence of `init` instructions when Y register are killed before
a function calls.