stdlib: Add option 'quote_singleton_atom_types' to erl_pp
The new option can be used by Dialyzer, which has traditionally quoted
singleton atom types.

The OTP team has decided not to use the convention in OTP
documentation, at least not for now. That is why the option is needed.

EDoc does not use the new option.

Syntax Tools does not have a similar option.