beam_emu.c: Avoid triggering an assertion for the wrong reason
Before 2d2e78ad6e66 that introduced tail-recursive calls of BIFs, the
stack was guaranteed not to be empty when `erlang:raise/3` was called
from the `catch` block of a `try` (because the `try` had set up a
stack frame that would be deallocated after the `raise` call).

Now the stack can be empty, so the ASSERT() call in next_catch()
that checks that there is a continuation pointer at the top of
the stack may fail.

Move the ASSERT() call to after check for empty stack.

While at it, also add a comment of the reason for the assertion.