erts: Remove unsafe bs_get_binary2 optimization from loader
A load-time optimization assumed that match contexts had no further
uses when a bs_get_binary2 overwrote the match context's register,
and figured it would be safe to reuse the match context's memory
for the resulting binary.

This is no longer safe as of OTP 22, as a match context may be
reused after being passed to another function.