[snmp|agent|test] Timestamps and test manager
Added common (formated) timestamp function(s). Made use of
these in the verbosity module (for debug printouts) and in
the test suite(s).

I also *think* I found the cause for some if the test case
failures (timeouts).
For v3 (agent) test cases the test manager makes use of parts
of the agent code: snmp_framework_mib and snmp_user_based_sm_mib.
And since they store their data in snmpa_local_db, that also
needs to be running.
And this was the problem (I think). On some (slow) machines,
the snmpa_local_db process from the *previous* test case
might still be running when the we tried to start it. That meant
that no new snmpa_local_db was started. Instead the old one,
still running but terminating, was retain. For a while. Until
it actually finally stopped. So the next operation towards
snmpa_local_db, insert, simply hanged until the process
terminated.
This in combination with the fact that the packet server process,
which was started using proc_lib, previously called init_ack
before this init was actually done, could actually start and then
at a much later time hang because some operation timed out
(the packet server was hanging).

Yuckety yuck-yuck.